shell/window-tracker: Track ::window-added again
Switching from ::window-added to ::window-created broke extensions
that look up a window's app on ::window-added, as that signal is
emitted before ::window-created and the window is therefore not
tracked yet.

For the upcoming release we can expect extensions to adjust to the
change, but for stable branches it is better to unbreak extensions
by tracking ::window-added again.

This partially reverts commit 5674dee6a4d5539941e8c.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/2120>