screenshot: Assign names to several CSS expressions
This commit replaces a few of the screenshot UI CSS expressions with
named variables.

The shot-cast margin is not defined, but rather set to the value it ends
up equal to, given the panel padding and the capture button's
sizes + position, which would be a bit awkward to compute here. It ends
up slightly larger than the general panel padding due to the capture
button being larger than other elements, and due to having to
center-align the shot-cast container to the capture button.

The shot-cast container border radius is defined as 12px, then the panel
border radius is computed from the shot-cast border radius + margin,
then the type button border radius is computed from panel
border radius - padding.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/2103>