Osk-layouts: Add Austrian German layout
The 'at' layout is identical to 'de', but as it is implemented as
a different layout rather than a variant, we can only fall back
to 'us'.

To fix that, copy the existing German layout into a new Austrian
one (with the 'name' and 'locale' keys adjusted accordingly).

https://gitlab.gnome.org/GNOME/gnome-shell/-/issues/5372

Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/2274>